# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PageSchema_PatientAllergyRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PatientAllergyRead"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "items",
          "page"
        ],
        "title": "PageSchema[PatientAllergyRead]",
        "type": "object"
      },
      "PageSchema_PatientFlagRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PatientFlagRead"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "items",
          "page"
        ],
        "title": "PageSchema[PatientFlagRead]",
        "type": "object"
      },
...
      "PageSchema_PatientSummaryRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PatientSummaryRead"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "items",
          "page"
        ],
        "title": "PageSchema[PatientSummaryRead]",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "EntShifa API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/patients": {
      "get": {
        "operationId": "list_patients_api_v1_patients_get",
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 190,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "birth_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Birth Date"
            }
          },
          {
            "in": "query",
            "name": "sex",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 10,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sex"
            }
          },
          {
            "in": "query",
            "name": "flag_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 60,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Flag Code"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 40,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sort"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Clinic-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Clinic-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageSchema_PatientSummaryRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Patients",
        "tags": [
          "patients"
        ]
      },
      "post": {
        "operationId": "create_patient_api_v1_patients_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Clinic-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Clinic-Id"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Patient",
        "tags": [
          "patients"
        ]
      }
    },
    "/api/v1/patients/{patient_id}": {
      "get": {
        "operationId": "get_patient_api_v1_patients__patient_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "title": "Patient Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Clinic-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Clinic-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Patient",
        "tags": [
          "patients"
        ]
      },
      "patch": {
        "operationId": "update_patient_api_v1_patients__patient_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "title": "Patient Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Clinic-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Clinic-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Patient",
        "tags": [
          "patients"
        ]
      }
    },
    "/api/v1/patients/{patient_id}/allergies": {
      "get": {
        "operationId": "list_allergies_api_v1_patients__patient_id__allergies_get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "title": "Patient Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Clinic-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Clinic-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageSchema_PatientAllergyRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Allergies",
        "tags": [
          "patients"
        ]
      },
      "post": {
        "operationId": "add_allergy_api_v1_patients__patient_id__allergies_post",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "title": "Patient Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Clinic-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Clinic-Id"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientAllergyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientAllergyRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Allergy",
        "tags": [
          "patients"
        ]
      }
    },
    "/api/v1/patients/{patient_id}/flags": {
      "get": {
        "operationId": "list_flags_api_v1_patients__patient_id__flags_get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "title": "Patient Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "flag_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 60,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Flag Code"
            }
          },
          {
            "in": "query",
            "name": "active_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Active Only",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Clinic-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Clinic-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageSchema_PatientFlagRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Flags",
        "tags": [
          "patients"
        ]
      },
      "post": {
        "operationId": "add_flag_api_v1_patients__patient_id__flags_post",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "title": "Patient Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Clinic-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Clinic-Id"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientFlagCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientFlagRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Flag",
        "tags": [
          "patients"
        ]
      }
    },
    "/api/v1/patients/{patient_id}/flags/{flag_id}": {
      "patch": {
        "operationId": "update_flag_api_v1_patients__patient_id__flags__flag_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "title": "Patient Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "flag_id",
            "required": true,
            "schema": {
              "title": "Flag Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Clinic-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Clinic-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientFlagUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientFlagRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Flag",
        "tags": [
          "patients"
        ]
      }
    },
...
    "/api/v1/patients/{patient_id}/merge": {
      "post": {
        "operationId": "merge_patient_api_v1_patients__patient_id__merge_post",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "title": "Patient Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Clinic-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Clinic-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientMergeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientMergeRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Merge Patient",
        "tags": [
          "patients"
        ]
      }
    },
...
  }
}